# docs/swagger.yaml
basePath: /
definitions:
...
  api.InsertMaterialParams:
    properties:
      color:
        type: string
      name:
        type: string
      quantity:
        type: string
      remarks:
        type: string
      size:
        type: string
    type: object
  api.InsertOrderItemParams:
    properties:
      product:
        $ref: '#/definitions/api.InsertOrderProductParams'
      quantity:
        type: integer
      totalPrice:
        type: number
    type: object
  api.InsertOrderParams:
    properties:
      customerId:
        type: string
      deliveryDate:
        type: string
      orderDate:
        type: string
      orderItems:
        items:
          $ref: '#/definitions/api.InsertOrderItemParams'
        type: array
      paymentDate:
        type: string
      shippingAddress:
        type: string
      status:
        type: string
      totalAmount:
        type: number
    type: object
  api.InsertOrderProductParams:
    properties:
      sku:
        type: string
      unitPrice:
        type: number
    type: object
...
  api.UpdateMaterialParams:
    properties:
      color:
        type: string
      name:
        type: string
      quantity:
        type: string
      remarks:
        type: string
      size:
        type: string
    type: object
  api.UpdateOrderItemParams:
    properties:
      product:
        $ref: '#/definitions/api.UpdateOrderProductParams'
      quantity:
        type: integer
      totalPrice:
        type: number
    type: object
  api.UpdateOrderParams:
    properties:
      customerId:
        type: string
      deliveryDate:
        type: string
      orderDate:
        type: string
      orderItems:
        items:
          $ref: '#/definitions/api.UpdateOrderItemParams'
        type: array
      paymentDate:
        type: string
      shippingAddress:
        type: string
      status:
        type: string
      totalAmount:
        type: number
    type: object
  api.UpdateOrderProductParams:
    properties:
      sku:
        type: string
      unitPrice:
        type: number
    type: object
...
  fiber.Map:
    additionalProperties: true
    type: object
  types.Material:
    properties:
      color:
        type: string
      id:
        type: string
      name:
        type: string
      quantity:
        type: string
      remarks:
        type: string
      size:
        type: string
    type: object
  types.Order:
    properties:
      customerId:
        type: string
      deliveryDate:
        type: string
      id:
        type: string
      orderDate:
        type: string
      orderItems:
        items:
          $ref: '#/definitions/types.OrderItem'
        type: array
      paymentDate:
        type: string
      shippingAddress:
        type: string
      status:
        type: string
      totalAmount:
        type: number
    type: object
  types.OrderItem:
    properties:
      product:
        $ref: '#/definitions/types.OrderProduct'
      quantity:
        type: integer
      totalPrice:
        type: number
    type: object
  types.OrderProduct:
    properties:
      sku:
        type: string
      unitPrice:
        type: number
    type: object
...
host: localhost:8080
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: This is a sample swagger for Fiber
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fiber Example API
  version: "1.0"
paths:
...
  /material:
    get:
      description: Get a list of materials based on the provided filters.
      parameters:
      - description: Material ID (optional)
        in: query
        name: id
        type: string
      - description: Material name (optional)
        in: query
        name: name
        type: string
      - description: Material color (optional)
        in: query
        name: color
        type: string
      - description: Material size (optional)
        in: query
        name: size
        type: string
      - description: Material quantity (optional)
        in: query
        name: quantity
        type: string
      - description: Material remarks (optional)
        in: query
        name: remarks
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.Material'
            type: array
      summary: Get materials
      tags:
      - Material
    post:
      consumes:
      - application/json
      description: Insert a new material into the system.
      parameters:
      - description: Material details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.InsertMaterialParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Material'
      summary: Insert material
      tags:
      - Material
  /material/{id}:
    delete:
      description: Delete a material from the system.
      parameters:
      - description: Material ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Delete material
      tags:
      - Material
    patch:
      consumes:
      - application/json
      description: Update an existing material in the system.
      parameters:
      - description: Material ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated material details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.UpdateMaterialParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Update material
      tags:
      - Material
...
  /order:
    get:
      description: Retrieves a list of orders based on query parameters.
      parameters:
      - description: Order ID
        in: query
        name: id
        type: string
      - description: Customer ID
        in: query
        name: customerId
        type: string
      - description: Order status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.Order'
            type: array
      summary: Get orders
      tags:
      - Order
    post:
      consumes:
      - application/json
      description: Inserts a new order.
      parameters:
      - description: Order information
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/api.InsertOrderParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Order'
      summary: Insert order
      tags:
      - Order
  /order/{id}:
    delete:
      description: Deletes an order by ID.
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Delete order
      tags:
      - Order
    patch:
      consumes:
      - application/json
      description: Update an existing order in the system.
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated order details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.UpdateOrderParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Update order
      tags:
      - Order
...
swagger: "2.0"